feat: Modified column names as per the new excel provided

+++ src/main/java/uk/gov/hmcts/reform/cwrdapi/service/impl/ExcelAdaptorServiceImpl.java
@@ -219,7 +219,7 @@ public class ExcelAdaptorServiceImpl implements ExcelAdaptorService {
             case BOOLEAN:
                 return cell.getBooleanCellValue();
             case NUMERIC:
-                return cell.getNumericCellValue();
+                return (int) cell.getNumericCellValue();
             case STRING:
                 return cell.getStringCellValue();
             default:
